// src/shared/constants/enum/stepCompany.enum.ts
export enum CompanyStepEnum {
  WORKSPACE = 'WORKSPACE',
  EMPLOYEE = 'EMPLOYEE',
  MANAGEMENT = 'MANAGEMENT',
  HOMO_GROUP = 'HOMO_GROUP',
  HIERARCHY = 'HIERARCHY',
  RISK_GROUP = 'RISK_GROUP',
  RISKS = 'RISKS',
  PROFESSIONALS = 'PROFESSIONALS',
  EXAMS = 'EXAMS',
  USERS = 'USERS',
  NONE = 'NONE',
}